Move IPC helpers from tools to utils
